Move mypyc.cstring to mypyc.codegen.cstring
