+++ mypyc/codegen/cstring.py
